Re order components.
